# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Merchant Platform API
  version: 0.2.0
  description: Tenant-scoped invoice, payment reconciliation, audit, and operations API.
servers:
  - url: http://localhost:8080
    description: Local
security:
  - PlatformKey: []
    ActorToken: []
paths:
...
  /v1/invoices:
    get:
      operationId: listInvoices
      parameters:
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
      responses:
        "200":
          description: Invoices visible to the actor merchant hierarchy
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Invoice" }
    post:
      operationId: createInvoice
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateInvoice" }
      responses:
        "201":
          description: Invoice issued; identical replay returns the same representation
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Invoice" }
        "400": { description: Invalid money, item, or idempotency input }
        "403": { description: Store is outside the actor merchant hierarchy }
        "409": { description: Idempotency key is bound to a different request }
...
  /v1/payments:
    get:
      operationId: listPayments
      parameters:
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 50 }
      responses:
        "200":
          description: Payments visible to the actor merchant hierarchy
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Payment" }
  /v1/payments/webhooks/{provider}:
    post:
      security: []
      operationId: acceptSignedPaymentWebhook
      parameters:
        - name: provider
          in: path
          required: true
          schema: { type: string, pattern: "^[a-z0-9-]{2,40}$" }
        - name: X-Payment-Timestamp
          in: header
          required: true
          schema: { type: string, pattern: "^[0-9]+$" }
        - name: X-Payment-Signature
          in: header
          required: true
          schema: { type: string, pattern: "^[a-fA-F0-9]{64}$" }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PaymentWebhook" }
      responses:
        "201": { description: Financial event, audit record, and outbox event committed atomically }
        "400": { description: Invalid callback schema or event-time window }
        "401": { description: Signature verification failed }
        "404": { description: Provider, invoice, or captured payment not found }
        "409": { description: Replay conflict, state conflict, over-capture, or over-refund }
...
components:
  securitySchemes:
    PlatformKey:
      type: apiKey
      in: header
      name: X-Platform-Api-Key
    ActorToken:
      type: http
      scheme: bearer
      bearerFormat: OIDC JWT
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string, minLength: 16, maxLength: 128, pattern: "^[A-Za-z0-9][A-Za-z0-9._:-]+$" }
...
  schemas:
...
    CreateInvoice:
      type: object
      additionalProperties: false
      required: [storeId, customerName, customerContact, items]
      properties:
        storeId: { type: string, format: uuid }
        customerName: { type: string, maxLength: 160 }
        customerContact: { type: string, format: email, maxLength: 254 }
        currency: { type: string, enum: [INR], default: INR }
        discountMinor: { type: integer, minimum: 0, maximum: 100000000, default: 0 }
        taxRateBasisPoints: { type: integer, minimum: 0, maximum: 10000, default: 0 }
        items:
          type: array
          minItems: 1
          maxItems: 100
          items:
            type: object
            additionalProperties: false
            required: [description, quantity, unitPriceMinor]
            properties:
              description: { type: string, maxLength: 240 }
              quantity: { type: integer, minimum: 1, maximum: 10000 }
              unitPriceMinor: { type: integer, minimum: 0, maximum: 100000000 }
    Invoice:
      type: object
      required: [id, storeId, status, currency, subtotalMinor, discountMinor, taxMinor, totalMinor, issuedAt, version]
      properties:
        id: { type: string, format: uuid }
        storeId: { type: string, format: uuid }
        status: { type: string, enum: [DRAFT, ISSUED, PARTIALLY_PAID, PAID, PARTIALLY_REFUNDED, REFUNDED, CANCELLED] }
        currency: { type: string, enum: [INR] }
        subtotalMinor: { type: integer }
        discountMinor: { type: integer }
        taxMinor: { type: integer }
        totalMinor: { type: integer }
        issuedAt: { type: string, format: date-time }
        version: { type: integer, minimum: 1 }
    PaymentWebhook:
      type: object
      additionalProperties: false
      required: [providerEventId, eventType, providerPaymentId, invoiceId, amountMinor, currency, occurredAt]
      properties:
        providerEventId: { type: string, minLength: 8, maxLength: 120 }
        eventType: { type: string, enum: [payment.captured, payment.refunded] }
        providerPaymentId: { type: string, minLength: 8, maxLength: 120 }
        providerRefundId: { type: string, minLength: 8, maxLength: 120 }
        invoiceId: { type: string, format: uuid }
        amountMinor: { type: integer, minimum: 1, maximum: 100000000 }
        currency: { type: string, enum: [INR] }
        occurredAt: { type: string, format: date-time }
    Payment:
      type: object
      required: [id, merchantId, invoiceId, provider, providerPaymentId, amountMinor, refundedMinor, currency, status, capturedAt, version]
      properties:
        id: { type: string, format: uuid }
        merchantId: { type: string, format: uuid }
        invoiceId: { type: string, format: uuid }
        provider: { type: string }
        providerPaymentId: { type: string }
        amountMinor: { type: integer, minimum: 1 }
        refundedMinor: { type: integer, minimum: 0 }
        currency: { type: string, enum: [INR] }
        status: { type: string, enum: [CAPTURED, PARTIALLY_REFUNDED, REFUNDED] }
        capturedAt: { type: string, format: date-time }
        version: { type: integer, minimum: 1 }
...